Initial support for clang-tidy
First changes to support clang-tidy through icecream. The current revision works but
misses some critial features. Running clang-tidy through icecream will work if any of the
include path are part of the command line (-I...). The only supported options
are --checks, --warnings-as-errors and --extra-args. ICECC_CARET_WORKAROUND also has to be
set to 0.

The following changes are required for icecc to work properly.

- If clang-tidy exports the -p option, we should find the command line in the compile_command.json
and extract the -I includes directly.
- .clang-tidy is not supported, maybe we could run a first pass with --dump-config and include
it in the command line (--config) when running it on the remote machine.
- Make sure ICECC_CARET_WORKAROUND is not needing when running clang-tidy
- Find clang++/clang using clang-tidy's path.
- Find a better way to handle clang-tidy's argument rather than adding them manually to the code.
- Delete the tmp file once the compilation is over.
- Find a solution for line directives not working (recompile locally on error or patch clang-tidy.?)
- How do we want to handle --export-fixes / --fix?

There are some limitations to this implementation that I doubt we will be able to address in the
near future. After running the preprocessor the macros / includes are
lost. So any check diagnosing issues with macros or includes will not work. The include path
will also be broken in the diagnosis.