Extract Include dir from compile_commands
If the argument -p is specified when running clang-tidy
Append the -I and --isystem to the preprocessor before
sending the file to the remote compiler.

The current implement is fragile, but does the trick. More work
will be needed in the future to make the parsing implementation
reliable on invalid jsons