Move compile_command analysis to argument parser
The analysis was previously done before running the preprocessor.
This is problematic since we have to extract the flags for clang-tidy
as well. (-std=c++XX) and this information is not available if the
preprocessor is the one extracting this info.

The info is now collected directly through arg.cpp and dispatched
to the preprocessor / remote node depending on the flag they need.

There seems to be 1 minor issue with the current implementation,
clang-tidy is somehow able to extract some compile_commands from
files that are not part of the compile_command.json file.
I'm not exactly sure how clang-tidy is able to do that, but I will
need to investigate this feature to properly support it in icecream as well.